requirements.txt
Flask
flask-talisman
python-dotenv
newsapi-python
